add custom string tables, fix old func
